requirements.txt
Django==4.2.1
Pillow==9.5.0
asgiref==3.6.0
autopep8==2.0.2
black==23.3.0
coreapi==2.3.3
crispy-bootstrap5==0.7
django-allauth==0.54.0
django-crispy-forms==2.0
django-debug-toolbar==4.1.0
django-filter==23.2
djangorestframework-simplejwt==5.2.2
djangorestframework==3.14.0
drf-spectacular-sidecar==2023.5.1
drf-spectacular==0.26.2
environs==9.5.0
mysqlclient==2.1.1
psycopg2-binary==2.9.6
pudb==2022.1.3
python-dotenv==1.0.0
sqlparse==0.4.4
whitenoise==6.4.0
